+++ src/embeds/amulets.js
@@ -2,8 +2,8 @@
 
 import type { EmbedProps } from './bootstrap';
 
-import Skills from './components/Skills';
+import Amulets from './components/Amulets';
 
 export default function (element: HTMLElement, ids: Array<number>) {
-  return (props: EmbedProps) => <Skills {...props} ids={ids} />;
+  return (props: EmbedProps) => <Amulets {...props} ids={ids} />;
 }