+++ src/embeds/character.js
@@ -1,9 +1,9 @@
 // @flow
 
-// TODO: Remove character.js and replace with this to use in new system.
+import type { EmbedProps } from './bootstrap';
 import Character from './components/Character';
 
 export default function (element: HTMLElement) {
-  const name = element.getAttribute('data-armory-name');
-  return () => <Character name={name} />;
+  const name = element.getAttribute('data-armory-name') || '';
+  return (props: EmbedProps) => <Character {...props} name={name} />;
 }