+++ src/embeds/index.js
@@ -1,5 +1,9 @@
 // @flow
 
+import * as ls from 'lib/localStorage';
+
+ls.reset();
+
 // $FlowFixMe
 const load = require('promise?global!./bootstrap'); // eslint-disable-line import/no-webpack-loader-syntax