+++ src/index.js
@@ -1,5 +1,9 @@
 // @flow
 
+import * as ls from 'lib/localStorage';
+
+ls.reset();
+
 import ReactDOM from 'react-dom';
 import Router from './Router';
 import Base from './Base';